Begin moving to a more abstract style of UGens
This work is towards splitting the visual computations into a number of
shader programs rather than generating one large fragment shader from
the entire node tree.

The purpose is to support texture operations on the output of any UGen.
For example, it was previously only possible to rotate a texture and
this mapped closely to the GLSL implementation -- separate
transformations on texture coordinates which are passed as an argument
to the texture rendering UGen.

In sclang, to play a video (with ID ~vidID) normally, you would write

    PlayVid.fr(~vidID)

but to rotate that video, the use of a different video playback UGen is
required.
    At the GLSL level, PlayVid takes a sampler2D and outputs a vec4 for
each fragment. Rotation is only possible when initially rendering a
texture with rotated texture coordinates.
    If you wanted to rotate the video with a sine wave LFO controlling
the angle it is rotated, you would have to use a different UGen and
manually transform the texture coordinates:

    ShowVidTex.fr(~vidID, Rotate.fr(TexCoords.fr, SinOsc.kr(0.1)))

It is a much nicer abstraction to be able to write:

    Rotate.fr(PlayVid.fr(~vidID), SinOsc.kr(0.1))

There will likely be a negative performance impact of this approach
which will need to be measured but it significantly improves the user
interface. Reduced user-facing complexity at the expense of a more
complex underlying implementation.