Use rio over default Prelude + various changes
This commit includes a bunch of changes which should probably be
separated into numerous separate commits but the time it'll take to do
that isn't worth it. Some changes are incomplete and more significant
changes will follow.

The motivation behind using rio over the default Prelude is that both
the cognitive and code-level complexity associated with making changes
and adding new features was getting out of hand. Each function was
taking far too many arguments when most or all of the arguments were
really just parts of the program state.